Support for posting to an elog.
We post a PNG with a comment and link to a URL specified in the
viewerVars. On the server side, the endpoint receives a multipart form
with a "comment", "link" form attributes and a "snapshot" file
attribute.

For example, in Flask terms, on the endpoint, we get
request.form["comment"]
request.form["link"]
request.files.getlist("snapshot")