Support for relative dates in the URL.
For example, "from=-1d" should get a days worth of data.
Another example, "from=-1w&to=-5d" should get two days worth of data
from a week ago to 5 days ago.
We use moment.js add/subtract shorthand for the characters suppported.
So, smhdwMQy should work.
This now determines the bin size when parsing the URL.
So, we support a binSize parameter and also add it to the generated URL.